refactor(wallet): Share nine more tabs onto the seams
LightningTab, IssuedTab, GroupsTab, JsonViewer, AliasedDIDs, PollTab, ImageTab,
FileTab and VaultTab. The extension takes react-wallet's design with them.

Three more host differences turned out to be derivable rather than injectable,
which is the cheapest kind:

- dark mode comes from the MUI theme's palette mode, since both wallets set it
  on the ThemeProvider they already wrap everything in. No seam, like
  isTabletUp before it.
- the node URL is now on the wallet context. FileTab and ImageTab were each
  reading it out of localStorage through app constants; the host already
  resolves it for WalletProvider, so it is published there instead.
- scanning an alias QR joins scanQr as a platform capability, so AliasedDIDs
  asks for a camera rather than importing one.

Pairs 21 -> 12.

Refs #894
