+++ apps/browser-extension/src/components/AssetsTab.tsx
@@ -3,10 +3,10 @@ import { Box, Tab } from "@mui/material";
 import {TabContext, TabList, TabPanel} from "@mui/lab";
 import {ControlPointDuplicate, Description, Groups, Image, Lock, Schema} from "@mui/icons-material";
 import { SchemaTab } from "@didcid/wallet-ui";
-import ImageTab from "./ImageTab";
-import FileTab from "./FileTab";
-import GroupsTab from "./GroupsTab";
-import VaultTab from "./VaultTab";
+import { ImageTab } from "@didcid/wallet-ui";
+import { FileTab } from "@didcid/wallet-ui";
+import { GroupsTab } from "@didcid/wallet-ui";
+import { VaultTab } from "@didcid/wallet-ui";
 import { CloneAssetTab } from "@didcid/wallet-ui";
 
 function AssetsTab({ subTab }: {subTab: string}) {
