+++ apps/browser-extension/src/components/BrowserContent.tsx
@@ -20,15 +20,15 @@ import WalletTab from "./WalletTab";
 import SettingsTab from "./SettingsTab";
 import IdentitiesTab from "./IdentitiesTab";
 import BrowserHeader from "./BrowserHeader";
-import JsonViewer from "./JsonViewer";
+import { JsonViewer } from "@didcid/wallet-ui";
 import { useWalletContext } from "@didcid/wallet-ui";
 import { useUIContext } from "../contexts/UIContext";
 import { useVariablesContext } from "@didcid/wallet-ui";
-import AliasedDIDs from "./AliasedDIDs";
+import { AliasedDIDs } from "@didcid/wallet-ui";
 import AssetsTab from "./AssetsTab";
 import DmailTab from "./DmailTab";
-import LightningTab from "./LightningTab";
-import PollTab from "./PollTab";
+import { LightningTab } from "@didcid/wallet-ui";
+import { PollTab } from "@didcid/wallet-ui";
 import AuthTab from "./AuthTab";
 import { PropertiesTab } from "@didcid/wallet-ui";
