+++ apps/browser-extension/src/components/HeldTab.tsx
@@ -5,7 +5,7 @@ import { useWalletContext } from "@didcid/wallet-ui";
 import { useSnackbar } from "@didcid/wallet-ui";
 import { useVariablesContext } from "@didcid/wallet-ui";
 import { useUIContext } from "../contexts/UIContext";
-import JsonViewer from "./JsonViewer";
+import { JsonViewer } from "@didcid/wallet-ui";
 import { requestBrowserRefresh } from "../utils/utils";
 import { DisplayDID } from "@didcid/wallet-ui";
