+++ apps/react-wallet/src/BrowserContent.tsx
@@ -37,19 +37,19 @@ import WalletTab from "./components/WalletTab";
 import SettingsTab from "./components/SettingsTab";
 import IdentitiesTab from "./components/IdentitiesTab";
 import BrowserHeader from "./components/BrowserHeader";
-import JsonViewer from "./components/JsonViewer";
+import { JsonViewer } from "@didcid/wallet-ui";
 import { PageHeader } from "@didcid/wallet-ui";
 import { useVariablesContext } from "@didcid/wallet-ui";
 import { useUIContext } from "./contexts/UIContext";
 import { useThemeContext } from "./contexts/ContextProviders";
 import { useWalletContext } from "@didcid/wallet-ui";
 import { useSafeArea } from "./contexts/SafeAreaContext";
-import AliasedDIDs from "./components/AliasedDIDs";
+import { AliasedDIDs } from "@didcid/wallet-ui";
 import AssetsTab from "./components/AssetsTab";
 import DmailTab from "./components/DmailTab";
-import LightningTab from "./components/LightningTab";
+import { LightningTab } from "@didcid/wallet-ui";
 import DidCommTab from "./components/DidCommTab";
-import PollTab from "./components/PollTab";
+import { PollTab } from "@didcid/wallet-ui";
 import AuthTab from "./components/AuthTab";
 import { PropertiesTab } from "@didcid/wallet-ui";
