+++ apps/react-wallet/src/components/CredentialsTab.tsx
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from "react";
 import { Box, Stack, Tab, Tabs } from "@mui/material";
 import { TabContext, TabPanel } from "@mui/lab";
 import { IssueTab } from "@didcid/wallet-ui";
-import IssuedTab from "./IssuedTab";
+import { IssuedTab } from "@didcid/wallet-ui";
 import HeldTab from "./HeldTab";
 import { PageHeader } from "@didcid/wallet-ui";
