--- apps/react-wallet/src/components/IssuedTab.tsx
@@ -1,198 +1,0 @@
-import { useState } from "react";
-import {
-    Box,
-    Button,
-    TextField,
-} from "@mui/material";
-import { useWalletContext } from "@didcid/wallet-ui";
-import { useVariablesContext } from "@didcid/wallet-ui";
-import { useUIContext } from "../contexts/UIContext";
-import { WarningModal } from "@didcid/wallet-ui";
-import JsonViewer from "./JsonViewer";
-import { DisplayDID } from "@didcid/wallet-ui";
-import {useSnackbar} from "@didcid/wallet-ui";
-
-function IssuedTab() {
-    const { keymaster } = useWalletContext();
-    const { setError } = useSnackbar();
-    const {
-        issuedEdit,
-        issuedList,
-        issuedString,
-        issuedStringOriginal,
-        selectedIssued,
-        setIssuedEdit,
-        setIssuedList,
-        setIssuedString,
-        setIssuedStringOriginal,
-        setSelectedIssued,
-    } = useVariablesContext();
-    const {
-        setOpenBrowser
-    } = useUIContext();
-    const [open, setOpen] = useState<boolean>(false);
-    const [revokeDID, setRevokeDID] = useState<string>("");
-
-    async function resolveIssued(did: string) {
-        setOpenBrowser({
-            did,
-            tab: "credentials",
-            subTab: "issued",
-        });
-        setSelectedIssued(did);
-        setIssuedEdit(false);
-    }
-
-    async function decryptIssued(did: string) {
-        if (!keymaster) {
-            return;
-        }
-        try {
-            const doc = await keymaster.getCredential(did);
-            setSelectedIssued(did);
-            const issued = JSON.stringify(doc, null, 4);
-            setIssuedStringOriginal(issued);
-            setIssuedString(issued);
-            setIssuedEdit(true);
-        } catch (error: any) {
-            setError(error);
-        }
-    }
-
-    async function updateIssued(did: string) {
-        if (!keymaster) {
-            return;
-        }
-        try {
-            const credential = JSON.parse(issuedString);
-            await keymaster.updateCredential(did, credential);
-            await decryptIssued(did);
-        } catch (error: any) {
-            setError(error);
-        }
-    }
-
-    async function handleRevokeConfirm() {
-        if (!keymaster) {
-            return;
-        }
-        try {
-            await keymaster.revokeCredential(revokeDID);
-            const newIssuedList = issuedList.filter((item) => item !== revokeDID);
-            setIssuedList(newIssuedList);
-            if (revokeDID === selectedIssued) {
-                setSelectedIssued("");
-                setIssuedEdit(false);
-                setIssuedString("");
-                setIssuedStringOriginal("");
-                setOpenBrowser({
-                    tab: "credentials",
-                    subTab: "issued",
-                });
-            }
-        } catch (error: any) {
-            setError(error);
-        }
-
-        setOpen(false);
-        setRevokeDID("");
-    }
-
-    const handleRevokeClose = () => {
-        setOpen(false);
-    };
-
-    const handleRevokeOpen = () => {
-        setOpen(true);
-    };
-
-    return (
-        <Box>
-            <WarningModal
-                title="Revoke Credential"
-                warningText="Are you sure you want to revoke the credential?"
-                isOpen={open}
-                onClose={handleRevokeClose}
-                onSubmit={handleRevokeConfirm}
-            />
-
-            <Box sx={{ height: 300, overflowY: 'auto', mb: 2 }}>
-                {issuedList && issuedList.map((did, index) => (
-                    <Box
-                        key={index}
-                        display="flex"
-                        flexDirection="column"
-                        alignItems="stretch"
-                        sx={{ mb: 2 }}
-                    >
-                        <DisplayDID did={did} />
-                        <Box className="flex-row">
-                            <Button
-                                variant="contained"
-                                color="primary"
-                                className="button large regular"
-                                onClick={() => resolveIssued(did)}
-                            >
-                                Resolve
-                            </Button>
-                            <Button
-                                variant="contained"
-                                color="primary"
-                                className="button large regular"
-                                onClick={() => decryptIssued(did)}
-                            >
-                                Decrypt
-                            </Button>
-                            <Button
-                                variant="contained"
-                                color="primary"
-                                className="button large regular"
-                                onClick={() => updateIssued(did)}
-                                disabled={did !== selectedIssued || !issuedEdit || issuedString === issuedStringOriginal}
-                            >
-                                Update
-                            </Button>
-                            <Button
-                                variant="contained"
-                                color="primary"
-                                className="button large regular"
-                                onClick={() => {
-                                    setRevokeDID(did);
-                                    handleRevokeOpen();
-                                }}
-                            >
-                                Revoke
-                            </Button>
-                        </Box>
-                    </Box>
-                ))}
-            </Box>
-            {selectedIssued &&
-                <Box>
-                    {(issuedEdit && issuedString) ? (
-                        <TextField
-                            value={issuedString}
-                            onChange={(e) => setIssuedString(e.target.value)}
-                            multiline
-                            rows={20}
-                            fullWidth
-                            variant="outlined"
-                            slotProps={{
-                                input: {
-                                    style: {
-                                        fontSize: "1em",
-                                        fontFamily: "Courier, monospace",
-                                    },
-                                }
-                            }}
-                        />
-                    ) : (
-                        <JsonViewer browserTab="credentials" browserSubTab="issued" />
-                    )}
-                </Box>
-            }
-        </Box>
-    );
-}
-
-export default IssuedTab;
