+++ apps/react-wallet/vite.config.ts
@@ -16,7 +16,7 @@ export default defineConfig({
         // @didcid/wallet-ui is consumed as source from outside this app, so
         // its imports would otherwise resolve a second copy of React and MUI
         // from the repo root. Two MUI instances means two emotion caches.
-        dedupe: ["@uiw/react-json-view", "@mui/material", "@mui/icons-material", "@emotion/react", "@emotion/styled", "react", "react-dom"],
+        dedupe: ["qrcode.react", "@uiw/react-json-view", "@mui/material", "@mui/icons-material", "@emotion/react", "@emotion/styled", "react", "react-dom"],
         alias: {
             "@didcid/wallet-ui": path.resolve(__dirname, "../../packages/wallet-ui/src/index.ts"),
             "@didcid/cipher/passphrase": path.resolve(__dirname, "../../packages/cipher/dist/esm/passphrase.js"),
