+++ packages/wallet-ui/src/index.ts
@@ -34,3 +34,13 @@ export { jsonViewTheme } from './components/layout/jsonViewTheme';
 export { default as IssueTab } from './components/IssueTab';
 export { default as CloneAssetTab } from './components/CloneAssetTab';
 export { default as SchemaTab } from './components/SchemaTab';
+export { default as LightningTab } from './components/LightningTab';
+export { default as IssuedTab } from './components/IssuedTab';
+export { default as GroupsTab } from './components/GroupsTab';
+export { default as JsonViewer } from './components/JsonViewer';
+export { useIsDarkMode } from './hooks/useIsTabletUp';
+export { default as AliasedDIDs } from './components/AliasedDIDs';
+export { default as PollTab } from './components/PollTab';
+export { default as ImageTab } from './components/ImageTab';
+export { default as FileTab } from './components/FileTab';
+export { default as VaultTab } from './components/VaultTab';
